react-toast

+++ src/Component/Mail/TrashMail.jsx
@@ -1,5 +1,10 @@
 import React, { useEffect, useState } from "react";
 import { Button, Card, ListGroup } from "react-bootstrap";
+import Loader from "../Layout/Loader";
+
+import { ToastContainer, toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
+import Swal from "sweetalert2";
 
 const TrashMail = () => {
   const mail = localStorage.getItem("email");
@@ -27,7 +32,7 @@ const TrashMail = () => {
       }
       setloader(false);
     } catch (err) {
-      console.log(err);
+      toast.error(err.message);
       setloader(false);
       setError(true);
     }
@@ -36,35 +41,64 @@ const TrashMail = () => {
   console.log(arr);
 
   const DeleteHandler = async (id) => {
-    try {
-      const res = await fetch(
-        `https://mail-box-ea204-default-rtdb.firebaseio.com/${email}trashMail/${id}.json`,
-        {
-          method: "DELETE",
-          headers: {
-            "Content-Type": "application/json",
-          },
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        try {
+          const res = await fetch(
+            `https://mail-box-ea204-default-rtdb.firebaseio.com/${email}trashMail/${id}.json`,
+            {
+              method: "DELETE",
+              headers: {
+                "Content-Type": "application/json",
+              },
+            }
+          );
+          let data = await res;
+          console.log(data);
+          setArr(arr.filter((i) => i[0] !== id));
+          // toast.success("Deleted");
+          Swal.fire("Deleted!", "Your file has been deleted.", "success");
+        } catch (error) {
+          toast.error("Something went wrong");
         }
-      );
-      let data = await res;
-      console.log(data);
-      setArr(arr.filter((i) => i[0] !== id));
-    } catch (error) {
-      alert(error.message);
-    }
+      }
+    });
   };
 
   useEffect(() => {
     getTrash();
-  }, [email]);
+  }, []);
 
   return (
-    <div>
-      {" "}
+    <>
+      <ToastContainer
+        position="top-right"
+        autoClose={3000}
+        hideProgressBar={false}
+        newestOnTop={false}
+        closeOnClick
+        rtl={false}
+        pauseOnFocusLoss
+        draggable
+        pauseOnHover
+        theme="dark"
+      />{" "}
       <Card className="scroll" bg="secondary">
         <h2 style={{ textAlign: "center" }}>Trashbox</h2>
         {error && arr.length === 0 && <h2>Something went wrong!</h2>}
-        {loader && <h5>Loading....</h5>}
+        {loader && (
+          <center>
+            <Loader />
+          </center>
+        )}
         <ListGroup>
           {!loader &&
             !error &&
@@ -113,8 +147,58 @@ const TrashMail = () => {
             })}
         </ListGroup>
       </Card>
-    </div>
+    </>
   );
 };
 
-export default TrashMail;
+export default React.memo(TrashMail);
+
+// 3.output:- a b e d c
+
+// 4.setTimeout doesn't return a promise so we can't await it. we can wrap setTimeout in a function that returns promise.
+
+// async function fun1(){
+
+// console.log('a');
+
+// console.log('b');
+
+// const fun2 = () =>{
+
+//  return new Promise((resolve,reject)=>{
+
+//   setTimeout(() => {
+
+//    console.log('c');
+
+//    resolve()
+
+//   }, 1000)
+
+//  })}
+
+// const fun3 = ()=>{
+
+//  return new Promise((res, rej)=>{
+
+//    setTimeout(() => {
+
+//    console.log('d');
+
+//    res()
+
+//   }, 0)
+
+//  })
+
+// }
+
+// await fun2()
+
+// await fun3()
+
+// console.log('e');
+
+// }
+
+// fun1()
